Edit display and capitalize constants
This version uses the value of the number 1
variable for the display value. In addition, it
uses an anonymous function instead of a named
function for adding a value to the display screen;
it appends the digits to the display value instead
of changing the value when entering multiple
digits; and it capitalizes all constants.